some changes

+++ Ex03.GarageLogic/AnswersValidator.cs
@@ -10,20 +10,19 @@ public class AnswersValidator
                 if (i_Questions[key].m_Type == typeof(int))
                 {
                     int value = int.Parse(i_Answers[key]);
-                    if (value < i_Questions[key].m_Min || value > i_Questions[key].m_Max)
+                    if (value < (int)i_Questions[key].m_Min || value > (int)i_Questions[key].m_Max)
                     {
-                        throw new ValueRangeException(i_Questions[key].m_Min, i_Questions[key].m_Max);
+                        throw new ValueRangeException((int)i_Questions[key].m_Min, (int)i_Questions[key].m_Max);
                     }
                 }
                 else
                 {
                     float value = float.Parse(i_Answers[key]);
-                    if (value < i_Questions[key].m_Min || value > i_Questions[key].m_Max)
+                    if (value < (float)i_Questions[key].m_Min || value > (float)i_Questions[key].m_Max)
                     {
-                        throw new ValueRangeException(i_Questions[key].m_Min, i_Questions[key].m_Max);
+                        throw new ValueRangeException((float)i_Questions[key].m_Min, (float)i_Questions[key].m_Max);
                     }
                 }
-
             }
         }
     }
